requirements.txt
pyspark
mlflow
kafka-python
boto3
pytest
